Move StatsCalculator from reporter to metrics
